Move more Owner.*Search class into the new module
